Set account on extended-join, simplify, fix docs
Now that account tracking exists, update extended-join to set account
data, too.  This was accidentally left out of the IRCv3 fix-up pull
request.

Simplify account tracking in 'processIrcEventAccount'.  No need to
check for '*' when that's used internally to represent logged-out.

Update documentation to mention 'extended-join' in WHOX and
IrcEventAccount handling.