icons: Add hicolor fallback for Qt < 5.5
Turns out that Qt only added support for split icon themes in 5.5,
so with older Qt versions our icon theme support doesn't work
properly.

Install hicolor icons for Qt < 5.5, as we already do for Qt 4.
Add a hicolor icon resource, too, and make use of it in the system
tray implementations, which is arguably the most important place
to have proper icon support.

This adds around 130 KiB to the binary size, but should save a lot
of headaches.